Fix: Correct D&D import in PDF Merge tool to resolve build error
Replaced placeholder 'react-dnd' imports with 'react-beautiful-dnd'
in `src/pages/tools/pdf/merge/Merge.tsx`. This resolves the Rollup
build failure caused by an incorrect module import.

The PDF Merge tool now correctly uses `DragDropContext` from
`react-beautiful-dnd` as intended for its drag-and-drop functionality.
