+++ src/components/tools/pdf-merge-wizard/steps/.gitkeep
@@ -1,0 +1,2 @@
+# This is a placeholder file to ensure the directory is created.
+# It can be deleted if not needed.
