+++ src/constants/tiers.ts
@@ -1,12 +1,14 @@
 // src/constants/tiers.ts
 
-// Assuming UserRole is imported or defined elsewhere (e.g., from AuthContext)
-// For now, let's define it here if not easily importable without creating circular deps,
-// or assume it will be imported from AuthContext.
-// If AuthContext is in src/context/AuthContext.tsx, the import would be:
-import type { UserRole } from '../context/AuthContext'; // Adjust path as needed
+// Define UserRole type if it's not globally available or imported from AuthContext
+// For simplicity here, assuming it might be defined elsewhere or should be added if not.
+// If AuthContext exports UserRole, it should be imported:
+// import { UserRole } from '../context/AuthContext';
 
-export const TIER_LIMITS: Record<UserRole, number> = {
+// If UserRole is not easily importable, define it locally for this file's scope:
+export type UserRole = "anonymous" | "free" | "premium" | "admin" | "superadmin";
+
+export const TIER_LIMITS = {
   anonymous: 20 * 1024 * 1024,  // 20 MB
   free:       20 * 1024 * 1024,  // 20 MB
   premium:   100 * 1024 * 1024,  // 100 MB
@@ -14,34 +16,15 @@ export const TIER_LIMITS: Record<UserRole, number> = {
   superadmin:100 * 1024 * 1024,  // 100 MB
 } as const;
 
-// Define which features are considered premium
-// This list should align with the feature matrix provided in the brief.
-// For example: 'OCR', 'ADVANCED_WATERMARK', 'EXTRA_COMPRESSION_MH' (Medium/High)
-const PREMIUM_FEATURES = [
-  'OCR',
-  'ADVANCED_WATERMARK',
-  'EXTRA_COMPRESSION' // Representing Medium/High compression
-] as const;
+export const PREMIUM_FEATURES = ['OCR', 'ADVANCED_WATERMARK'] as const;
 
-type PremiumFeature = typeof PREMIUM_FEATURES[number];
+// Ensure UserRole type is correctly sourced for this function.
+export function hasFeature(role: UserRole, feature: typeof PREMIUM_FEATURES[number] | string): boolean {
+  // Check if the feature is one of the defined premium features
+  const isPremiumFeature = (PREMIUM_FEATURES as readonly string[]).includes(feature);
 
-// Utility function to check if a user has access to a specific feature
-export function hasFeature(role: UserRole, feature: PremiumFeature | string): boolean {
-  // First, check if the feature string is one of the defined PremiumFeature types
-  const isListedPremiumFeature = (PREMIUM_FEATURES as readonly string[]).includes(feature);
-
-  if (isListedPremiumFeature) {
-    // If it's a listed premium feature, only premium+ roles have access
+  if (isPremiumFeature) {
     return ['premium', 'admin', 'superadmin'].includes(role);
   }
-  // If it's not in the PREMIUM_FEATURES list, assume it's a free/standard feature
-  return true;
+  return true; // Non-premium features are available to all
 }
-
-// Example of how it might be used for specific compression levels if needed:
-// export function hasCompressionLevelAccess(role: UserRole, level: 'low' | 'medium' | 'high'): boolean {
-//   if (level === 'medium' || level === 'high') {
-//     return hasFeature(role, 'EXTRA_COMPRESSION');
-//   }
-//   return true; // 'low' compression is free
-// }
